zlib: Remove gzgetc macro
There still is the gzgetc function (and also the gzgetc_ wrapper) which
can be used instead. In the original zlib.h, gzgetc is redefined as a macro
"for speed", but I don't think it's worth it to translate it to FB.

Translating it to FB would in fact not be trivial due to the use of -- and
++ operators. The only way to translate it would be as a private function,
and at that point we may aswell use the original gzgetc function from zlib.