raw: zlib: Rename things called OPAQUE to avoid conflict with wingdi.bi
The Windows API GDI header (wingdi) contains this:
	#define OPAQUE 2
which commonly causes problems with zlib.bi if that's included behind
windows.bi, because zlib uses "opaque" as identifier.

Of course the Windows headers contain more such #defines which could easily
trigger similar problems, so maybe one day we'll have to do this kind of
renaming for every header that might be used together with windows.bi,
not just in this case with zlib.

On the other hand, it's enough to change the order of #includes to work
around the issue, so renaming isn't really important, it's just a bit more
convenient for users. This fix up of zlib.bi was done in earlier versions
of FB (0.23) so it's good for backwards compatibility too.