raw: winapi: Translate headers as part of windows.h where needed
There's a lot of files that make up the Windows API headers, and it's not
clear which of them are intended to be included indirectly through
windows.h, or directly.

It definitely matters to the translation process though, because there are
lots of duplicate declarations which are guarded by #if checks. Since
fbfrog does preprocessing but does not necessarily complain about
incomplete input, it's best to include windows.h first everytime, because
that's the main entry point. Then at least we can be sure that from
windows.bi's point of view everything will work.

Of course that doesn't work with all headers; e.g. ntdef.h isn't supposed
to be used together with windows.h, or winsock2.h which is supposed to be
#included before windows.h, so it can override winsock.h...