winapi: Don't rename fields conflicting with global #defines
To match the C headers, the fields now have their original names.
The global #defines of the same name may change the identifiers, at least
if seen first.

When this happens, it will probably go unnoticed and not cause any
problems, as long as it affects both the declaration and any access to
the field.

Renaming the fields epxlicitly would require further fix ups in macros
containing accesses to these fields (because fbfrog doesn't update
identifiers in field accesses currently).