winapi: dpaddr: Fix up wstring literals
It looks like originally the Wine headers (which share code with MinGW-w64)
used
	#define DPNA_HEADER L"x-directplay:/"
but then that was changed to
	#define DPNA_HEADER (const WCHAR []){ 'x','-','d','i','r','e', ...
for the __GNUC__ case. Presumably that was done to allow using the headers
with a Linux gcc, where wchar_t and L"..." literals are different than for
Win32/Win64.

Of course for FB we don't really need that. And it couldn't be done at the
moment anyways - especially because fbfrog doesn't consider this at all
so far.