winapi: Use Extern "Windows" instead of Extern "C" for 64bit
Currently fbc still handles x86 callconvs on win64 as on win32 for the
purposes of type checking. I.e. the default is still "stdcall", and thus
with Extern "C" everything will be considered "cdecl" and incompatible to
FB's default, which matters for callback functions etc. that are written
in FB code outside Extern "C" blocks. Thus it's better to use Extern
"Windows" for stuff that also uses it on win32, because then the user's
code will work on both win32/win64.