winapi: sql*: Fix name conflicts without renaming
sql/sqlext provide plain functions. sqlucode provides *A and *W functions,
and #defines the plain names to *W under UNICODE.

Normally I'd fix this by renaming the plain functions in sql/sqlext to *A,
and then #defining the plain names to the *A versions under non-UNICODE,
but since the *A versions are already provided in sqlucode, the plain
declarations are now excluded from UNICODE in sql/sqlext (and will be
provided by sqlucode instead, as they would be in C).