freetype: Override ftoption.h to always get FT_Int64 & co
ftconfig.h checks whether "long" is 64bit, and if not, falls back to
checking for various compilers (to see whether it can use "long long" or
"__int64" etc.), but only if
	!defined __STDC__ || defined FT_CONFIG_OPTION_FORCE_INT64
However, fbfrog defines __STDC__. So then we have to pre-#define
FT_CONFIG_OPTION_FORCE_INT64. But that doesn't work either, because
ftoption.h #undefs it. (that's the default off setting)

So we have to modify ftoption.h to override this.