Regenerate with fbfrog 5383ffd1 (preserve all #defines/#undefs)
With this change, overridden #defines and #undefs are now preserved. They
are not part of a header files exposed API, strictly speaking, but on the
other hand there are cases of separate .h files where one overrides the
other's #defines, and sometimes that's not wanted, because it means the
#defines in the one .h file will be silently lost.

It's good that FB supports #undefs aswell, and errors on conflicting
#defines but allows redundant #defines, so this will in most cases not even
make a visible difference to the binding user.

Except that, since #defines aren't lost anymore, they may now be available,
if using one .bi file but not the other that overrides them.

An example where this matters is the winapi binding: winuser.bi #defines
lstrcpy, and strsafe.bi #undefs and re-#defines it. If using winuser.bi
alone then its lstrcpy should be available. If using strsafe.bi, it should
override lstrcpy.

Besides, this also exposes problems with the binding generation process,
e.g. the generation of SDL_config.h for the SDL headers... Fixing those
issues will only be good.