Update for fbfrog 1.9
The change to win/audevcod.bi happens because MinGW-w64 contains
two audevcod.h's - one in the normal winapi headers, one in the
direct-x headers, and now with the fbfrog changes, the main one
will be used (not the direct-x one). It doesn't matter though,
since they're effectively the same anyways.