Auto-build the needed fbfrog version
The idea is to associate the exact fbfrog version with the binding
generation scripts. Then it will no longer be necessary to enter the fbfrog
version/commit id into the commit messages here, and the headers are
easily reproducible at each version, no need to find the proper fbfrog
version anymore.