raw: Change NULL constants from cptr(any ptr, 0) to just 0
The point is that traditionally, NULL has just been a 0 integer in FB
headers, and that allows NULL to be passed to things like WPARAM/LPARAM
(winapi) directly without warning. This isn't even possible with the C
headers, but in FB it has worked like that since 10 years and now people
expect it.

http://sourceforge.net/p/fbc/bugs/779/