raw: Rename TRUE => CTRUE
Everywhere TRUE is defined as 1 it should be renamed to CTRUE to avoid
conflict with FB's built-in boolean TRUE, because that's -1.

Allegro 4 is an exception as it defines TRUE as -1 already.

Manual fix-ups are needed so that all C bindings with TRUE=1 will do
something like this:
	const CTRUE = 1
	#ifndef TRUE
		const TRUE = 1
	#endif
that allows CTRUE to be used in the presence of TRUE, but also allows
FB's built-in TRUE to be #undeffed to get the headers C-style TRUE.
Much like the fix-up done for winapi's BOOLEAN typedef.