winapi: Remove #inclib d3dx9d
In practice, programs have to be linked against one of many d3dx9* dlls,
for example:
	d3dx9_33
	d3dx9_39
	d3dx9_...
maybe plain d3dx9 (no d at the end) and d3dx9d (with d at the end) also are
possibilities sometimes (it probably depends on the installed DirectX
versions and such).

Anyways, it's clear that d3dx9d is not always the correct choice.
Furthermore, having that #inclib there means that linking silently works
and users aren't notified about being able to (or having to) choose.